refactor: move login_client to WsClient
